Only use one list of functions instead of two
